+++ tybotRoute/models/TiledeskChatbotUtil.js
@@ -480,6 +480,13 @@ class TiledeskChatbotUtil {
         return false;
     }
 
+    static isAudioMessage(message){
+        if (message && message.type && message.type === 'file' && message.metadata && message.metadata.src && message.metadata.type.includes('audio') ) {
+            return true;
+        }
+        return false;
+    }
+
     static lastUserMessageFrom(msg) {
         let message = {};
         message["senderFullname"] = msg["senderFullname"]; // ex. "Bot"
@@ -700,6 +707,14 @@ class TiledeskChatbotUtil {
                 await chatbot.addParameter(TiledeskChatbotConst.REQ_DEPARTMENT_NAME_KEY, message.attributes.departmentName);
             }
 
+            if (message && message.request && message.request.attributes && message.request.attributes.payload) {
+                if (!message.attributes) {
+                    message.attributes = {}
+                }
+                message.attributes.payload = { ...message.attributes.payload, ...message.request.attributes.payload}
+                if (chatbot.log) {console.log("FORCED SET message.attributes.payload:", JSON.stringify(message.attributes.payload))}
+                // if (projectId === "641864da99c1fb00131ba495") {console.log("641864da99c1fb00131ba495 > FORCED SET message.attributes.payload:", JSON.stringify(message.attributes.payload))}
+            }
             if (message.attributes) {
                 if (chatbot.log) {console.log("Ok message.attributes", JSON.stringify(message.attributes));}
                 // if (projectId === "641864da99c1fb00131ba495") {console.log("641864da99c1fb00131ba495 > Ok message.attributes", JSON.stringify(message.attributes));}
@@ -735,15 +750,7 @@ class TiledeskChatbotUtil {
                 }
             }
 
-            /** DEPRECATED */
-            if (message && message.request && message.request.attributes && message.request.attributes.payload) {
-                if (!message.attributes) {
-                    message.attributes = {}
-                }
-                message.attributes.payload = message.request.attributes.payload
-                if (chatbot.log) {console.log("FORCED SET message.attributes.payload:", JSON.stringify(message.attributes.payload))}
-                // if (projectId === "641864da99c1fb00131ba495") {console.log("641864da99c1fb00131ba495 > FORCED SET message.attributes.payload:", JSON.stringify(message.attributes.payload))}
-            }
+            
             
             
             const _bot = chatbot.bot; // aka FaqKB
@@ -926,8 +933,7 @@ class TiledeskChatbotUtil {
      * @param {string} requestId. Tiledesk chatbot/requestId parameters
      */
     getChatbotParameters(requestId, callback) {
-        // const jwt_token = this.fixToken(token);
-        const url = `${process.env.TYBOT_ENDPOINT}/ext/reserved/parameters/requests/${requestId}?all`;
+        const url = `${process.env.TILEBOT_ENDPOINT}/ext/reserved/parameters/requests/${requestId}?all`;
         const HTTPREQUEST = {
             url: url,
             headers: {