+++ tybotRoute/test/conversation-form-test.js
@@ -50,8 +50,9 @@ describe('Conversation1 - Form filling', async () => {
       console.log("Starting tilebot server...");
       tybot.startApp(
         {
-          // MONGODB_URI: process.env.mongoUrl,
+          // MONGODB_URI: process.env.MONGODB_URI,
           bots: bots_data,
+          TILEBOT_ENDPOINT: process.env.TILEBOT_ENDPOINT,
           API_ENDPOINT: process.env.API_ENDPOINT,
           REDIS_HOST: process.env.REDIS_HOST,
           REDIS_PORT: process.env.REDIS_PORT,
@@ -278,7 +279,7 @@ describe('Conversation1 - Form filling', async () => {
   });
 
   it('(intent-to-intent) /move_to => /target_intent', (done) => {
-    console.log("ALWAYS PASSES: (intent-to-intent) /move_to => /target_intent");
+    // console.log("ALWAYS PASSES: (intent-to-intent) /move_to => /target_intent");
     try {
       let listener;
       let endpointServer = express();
@@ -1007,8 +1008,7 @@ describe('Conversation1 - Form filling', async () => {
  * @param {string} token. User token
  */
 function sendMessageToBot(message, botId, token, callback) {
-  // const jwt_token = this.fixToken(token);
-  const url = `${process.env.TYBOT_ENDPOINT}/ext/${botId}`;
+  const url = `${process.env.TILEBOT_ENDPOINT}/ext/${botId}`;
   // console.log("sendMessageToBot URL", url);
   const HTTPREQUEST = {
     url: url,
@@ -1042,8 +1042,7 @@ function sendMessageToBot(message, botId, token, callback) {
  * @param {string} requestId. Tiledesk chatbot/requestId parameters
  */
 // function getChatbotParameters(requestId, callback) {
-//   // const jwt_token = this.fixToken(token);
-//   const url = `${process.env.TYBOT_ENDPOINT}/ext/parameters/requests/${requestId}?all`;
+//   const url = `${process.env.TILEBOT_ENDPOINT}/ext/parameters/requests/${requestId}?all`;
 //   const HTTPREQUEST = {
 //     url: url,
 //     headers: {