+++ tybotRoute/tiledeskChatbotPlugs/directives/DirConnectBlock.js
@@ -1,4 +1,3 @@
-//const { ExtApi } = require('../../ExtApi.js');
 let axios = require('axios');
 let https = require("https");
 const { v4: uuidv4 } = require('uuid');
@@ -63,14 +62,8 @@ class DirConnectBlock {
       "token": this.token
     }
     if (this.log) {console.log("move to intent message:", intent_command_request);}
-    let TILEBOT_ENDPOINT;
-    if (this.TILEBOT_ENDPOINT) {
-      TILEBOT_ENDPOINT = this.TILEBOT_ENDPOINT;
-    }
-    else {
-      TILEBOT_ENDPOINT = `${this.API_ENDPOINT}/modules/tilebot`
-    }
-    this.sendMessageToBot(TILEBOT_ENDPOINT, intent_command_request, botId, () => {
+
+    this.sendMessageToBot(this.TILEBOT_ENDPOINT, intent_command_request, botId, () => {
       callback();
     });
   }
@@ -114,9 +107,8 @@ class DirConnectBlock {
    * @param {string} botId. Tiledesk botId
    * @param {string} token. User token
    */
-  sendMessageToBot(CHATBOT_ENDPOINT, message, botId, callback) {
-    // const jwt_token = this.fixToken(token);
-    const url = `${CHATBOT_ENDPOINT}/ext/${botId}`;
+  sendMessageToBot(TILEBOT_ENDPOINT, message, botId, callback) {
+    const url = `${TILEBOT_ENDPOINT}/ext/${botId}`;
     // console.log("sendMessageToBot URL", url);
     const HTTPREQUEST = {
       url: url,