+++ tybotRoute/tiledeskChatbotPlugs/directives/DirMessageToBot.js
@@ -45,14 +45,8 @@ class DirMessageToBot {
       "token": this.token
     }
     if (this.log) {console.log("sending message:", JSON.stringify(outgoing_message));}
-    let TILEBOT_ENDPOINT;
-    if (this.TILEBOT_ENDPOINT) {
-      TILEBOT_ENDPOINT = this.TILEBOT_ENDPOINT;
-    }
-    else {
-      TILEBOT_ENDPOINT = `${this.API_ENDPOINT}/modules/tilebot`
-    }
-    this.sendMessageToBot(TILEBOT_ENDPOINT, outgoing_message, botId, () => {
+    
+    this.sendMessageToBot(this.TILEBOT_ENDPOINT, outgoing_message, botId, () => {
       // console.log("(DirMessageToBot) sendMessageToBot() req_body sent");
       callback(true);
     });
@@ -66,9 +60,8 @@ class DirMessageToBot {
    * @param {string} botId. Tiledesk botId
    * @param {string} token. User token
    */
-  sendMessageToBot(CHATBOT_ENDPOINT, message, botId, callback) {
-    // const jwt_token = this.fixToken(token);
-    const url = `${CHATBOT_ENDPOINT}/ext/${botId}`;
+  sendMessageToBot(TILEBOT_ENDPOINT, message, botId, callback) {
+    const url = `${TILEBOT_ENDPOINT}/ext/${botId}`;
     // console.log("sendMessageToBot URL", url);
     const HTTPREQUEST = {
       url: url,