+++ tybotRoute/tiledeskChatbotPlugs/directives/DirReplyV2.js
@@ -258,8 +258,19 @@ class DirReplyV2 {
       let userFlowAttributes = TiledeskChatbotUtil.userFlowAttributes(requestAttributes);
       if (this.log) { console.log("userFlowAttributes:", userFlowAttributes); }
       if (userFlowAttributes) {
-        message.attributes["flowAttributes"] = userFlowAttributes;
+        message.attributes["flowAttributes"] = {};
+        for (const [key, value] of Object.entries(userFlowAttributes)) {
+          try {
+            if(typeof value === 'string' && value.length <= 1000){
+              message.attributes["flowAttributes"][key] = value;
+            }
+          }
+          catch(err) {
+            console.error("An error occurred while JSON.parse(). Parsed value:" + value + " in allParametersStatic(). Error:", err);
+          }
+        }
       }
+      
     }
     // send!
     let cleanMessage = message;